+++ src/curious_pyapi/exceptions/exceptions.py
@@ -7,7 +7,7 @@ from httpx import codes, HTTPStatusError, RequestError, Response
 
 
 def allow_existing(
-    post: Callable[Any, Response], *, warn_if_existing: bool = True, **kwargs: Any
+    post: Callable[[], Response], *, warn_if_existing: bool = True, **kwargs: Any
 ) -> Response | None:
     """Post, but allow existing."""
     try:
@@ -18,22 +18,39 @@ def allow_existing(
             if getattr(_response, "status_code") == codes.BAD_REQUEST:
                 results = getattr(_response, "json")().get("result", [])
                 if results:
-                    if isinstance(results[0], dict) and results[0].get("message") == (
-                        "That email address is already associated "
-                        "with a Curious account."
-                    ):
+                    if isinstance(results[0], dict):
+                        _request = getattr(e, "request")
+                        warning: tuple[str, ...]
+                        message = results[0].get("message")
+                        match message:
+                            case (
+                                "That email address is already associated "
+                                "with a Curious account."
+                            ):
+                                warning = (
+                                    "%s is already associated with a Curious account.",
+                                    str(
+                                        json.loads(
+                                            getattr(_request, "content", json.dumps({}))
+                                        ).get("email", "That email address")
+                                    ),
+                                )
+                            case "Non-unique value.":
+                                warning = (
+                                    (
+                                        "A record with the same unique "
+                                        "value already exists."
+                                    ),
+                                )
+                            case _:
+                                raise
                         if warn_if_existing:
                             from ..utils.logging import get_logger  # noqa: PLC0415
 
                             logger = get_logger(__name__)
-                            _request = getattr(e, "request")
-                            logger.warning(
-                                "%s is already associated with a Curious account.",
-                                json.loads(
-                                    getattr(_request, "content", json.dumps({}))
-                                ).get("email", "That email address"),
-                            )
-        return None
+                            logger.warning(*warning)
+                        return None
+        raise
     return response